project/pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "project using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.102.0,<1.0.0",
    "google-generativeai>=0.8.4",
    "langchain>=0.3.20",
    "python-dotenv>=1.0.1",
    "streamlit>=1.43.0",
]

[project.scripts]
kickoff = "project.main:kickoff"
plot = "project.main:plot"
run_crew = "project.main2:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
